Support multi-line documents
Document contents is now an array of strings.  It's rendered by
including a <br> between each string.

This matches reasonably well with the contentEditable behavior in
Firefox.  Safari does some more fancy things with divs.

Reading the modified buffer into an array of strings: we keep text
nodes, and recurse into others.  In addition, we check for repeated <br>
tags (or <br> tags inside divs) and add an empty line so they will be
reconstituted.