Abort old change polls
Previously, we had no method to abort changelist longpolls.  This led to
their connections clogging the browser's list of max parallel
connections, making the app unresponsive.

This requires adding our own version of jquery.couch.js, using a patch
by Fedor Indutny (@indutny).

We keep track of the last change handler used, in order to stop it when
we start a new change.