Editable titles, save status icon
Added save status icon to top right corner of document view.  Icons are
located in _attachments/image/

Removed empty gravatar div from the "view" template.  It was taking up a
64x64 square, but not displaying properly.  Could have fixed it, but I
don't know that we needed a gravatar right there.

Made titles editable.  Disabled enter key in the title.  Doesn't
entirely solve the problem of multi-line titles, but it helps.  Also
prevented <br> from being saved in the title, and prevented backspace in
the title field when it is empty, saving the user from removing the
title field all together.