log out to homepage && the person not log in can't edit any documents.
I just know how to write commit changes. Sorry for meaningless message last time. Last time I make two changes. One is when we go to the home, the sidebar will become empty. Another is when we type into wrong password, there is an alert.
This time I make two changes. One is when we log out, it will turn to homepage and another is the person not log in can't edit any documents.

Today, I will import bootstrap.css to make it look good.